backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
pandas>=1.5.0
plotly>=5.0.0
openpyxl>=3.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
